🔨 chore: move back `genLLMConfig.ts`, add parm support for `genUserLLMConfig`
