+++ src/const/settings/llm.ts
@@ -1,8 +1,16 @@
 import { ModelProvider } from '@/libs/agent-runtime';
 
-import { genUserLLMConfig } from '@/utils/genLLMConfig'
+import { genUserLLMConfig } from '@/server/globalConfig/genLLMConfig'
 
-export const DEFAULT_LLM_CONFIG = genUserLLMConfig();
+export const DEFAULT_LLM_CONFIG = genUserLLMConfig({
+  ollama: {
+    enabled: true,
+    fetchOnClient: true,
+  },
+  openai: {
+    enabled: true,
+  },
+});
 
 export const DEFAULT_MODEL = 'gpt-4o-mini';
 export const DEFAULT_EMBEDDING_MODEL = 'text-embedding-3-small';
