+++ src/server/globalConfig/genLLMConfig.ts
@@ -48,17 +48,7 @@ export const genServerLLMConfig = () => {
   }, {} as Record<ModelProvider, any>);
 };
 
-export const genUserLLMConfig = (): UserModelProviderConfig => {
-  const specificConfig: Record<any, any> = {
-    ollama: {
-      enabled: true,
-      fetchOnClient: true,
-    },
-    openai: {
-      enabled: true,
-    },
-  };
-
+export const genUserLLMConfig = (specificConfig: Record<any, any>): UserModelProviderConfig => {
   return Object.keys(ModelProvider).reduce((config, providerKey) => {
     const provider = ModelProvider[providerKey as keyof typeof ModelProvider];
     const providerCard = ProviderCards[`${providerKey}ProviderCard` as keyof typeof ProviderCards] as ModelProviderCard;
@@ -73,3 +63,4 @@ export const genUserLLMConfig = (): UserModelProviderConfig => {
     return config;
   }, {} as UserModelProviderConfig);
 };
+
