+++ src/server/globalConfig/index.ts
@@ -8,7 +8,7 @@ import { GlobalServerConfig } from '@/types/serverConfig';
 
 import { parseAgentConfig } from './parseDefaultAgent';
 
-import { genServerLLMConfig } from '@/utils/genLLMConfig'
+import { genServerLLMConfig } from './genLLMConfig'
 
 export const getServerGlobalConfig = () => {
   const { ACCESS_CODES, DEFAULT_AGENT_CONFIG } = getAppConfig();
